requirements.txt
openai>=1.12.0
pandas>=2.0.0
openpyxl>=3.0.0
python-dotenv>=1.0.0
tqdm==4.66.5